Fixed a few details of method lookup in Objective-C symbols.  Now we find the correct method.
Unfortunately we don't get the superclass from the
runtime yet so the method doesn't import correctly
(and I added a check to make sure that doesn't hurt
us) but once we get that information right we will
report methods correctly to the parser as well.

Getting superclass information requires a common AST
context for all Objective-C runtime information,
meaning that the superclass and the subclass are in
the same AST context in all cases.  That is the next
thing that needs to be done here.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@146089 91177308-0d34-0410-b5e6-96231b3b80d8