<rdar://problem/11095005>
Fixed a performance regression when dynamic types are enable where we would ask a C++ type if it can possibly be dynamic. Previously we would force the type to complete itself and then anwwer the question definitively. Now we ask the type if it is already complete and only definitively answer the question for completed types and just say "yes" for non-complete C++ types. We also always now answer yes for Objective C classes and do not complete those types either.



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@153284 91177308-0d34-0410-b5e6-96231b3b80d8