Added a stress-tester for LLDB's disassembler. Right now it only works on Mac OS X, but other platforms would just need to add their own implementation of AddLLDBToSysPathOn*().
The stress-tester has two modes:

Used with --bytes N --random, the stress-tester
generates random instructions of length N and
runs them through the disassembler.  This is
suitable for architectures like Intel where it
is combinatorially infeasible to run through the
entire space of possible instructions.

Used with --bytes N and no arguments (or --start
S --stride T), the stress-tester tests the
disassembler with a monotonically increasing
sequence of instructions.

The --start and --stride arguments are intended
for use in multiprocessing environments.  Give
each core an ID from 0 .. T-1, pass the ID in as
the --start, and use T as the stride, and you
can launch one copy of the stress-tester on each
core you have available.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@154143 91177308-0d34-0410-b5e6-96231b3b80d8