Add support for function with stack frame checks added by the compiler; these functions will end in the sequence
  mov %rbp, %rsp
  ret
  call __stack_chk_fail

instead of the usual mov, ret.  The x86 assembly profiler only looked
for functions ending in 'ret' and added the Unwind row describing how to
set the CFA based on that -- the addition of the call insn (which is jumped
to earlier in the function body) threw off that inspection.

Resolves the need to "step" twice to get out of these functions when doing
source-level stepping.

<rdar://problem/11469705>



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@157454 91177308-0d34-0410-b5e6-96231b3b80d8