Extra safeguards to ensure that we never query the runtime if we have complete debug information for a class.
Also made the Objective-C language runtime return
NULL when asked for the complete debug information
(i.e., information from DWARF, not information from
the runtime) if that information isn't present.  It
used to return a non-authoritative version, which
made it hard for clients to determine whether
complete debug information was available.

<rdar://problem/12608895>


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@167299 91177308-0d34-0410-b5e6-96231b3b80d8