Setting TestInferiorChanged.py to expectedFailureLinux because of bugzilla 14662.
Note that the test actually fails before reaching the assertion described in that bug, but the failure is a trivial matter of adding OS-conditional expected stop reason strings (as noted by new comments in this patch).  However, with the stop reason strings changed, the test fails for the more substantial reason noted in bugzilla 14662.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@170619 91177308-0d34-0410-b5e6-96231b3b80d8