Add a new python file to help diagnose backtrace failures in the field. It will be installed in the LLDB.framework and can be loaded with
  (lldb) script import lldb.macosx

after which a "unwind-diagnose" command will be registered.  Select
the thread which has a bad backtrace and run this command -- a lot
of information about the stack frames, and an alternate backtrace
algorithm, will be used.  The information will often be sufficient
for a remote person to figure out why the backtrace failed.
<rdar://problem/13679300> 



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@180077 91177308-0d34-0410-b5e6-96231b3b80d8