elf-core: Remove SetTriple workaround
On FreeBSD I see that Arch, Vendor, OS, and Environment are unchanged
after this call (for x86_64, at least), and core debugging works without
it.

If we need to restore it we should be able to switch on
arch.GetTriple.getOS() instead of a compile-time #ifdef.

Thanks to Daniel Malea for testing on Linux.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@186862 91177308-0d34-0410-b5e6-96231b3b80d8