Added a 'jump' command, similar to GDBs.
This allows the PC to be directly changed to a different line.
It's similar to the example python script in examples/python/jump.py, except implemented as a builtin.

Also this version will track the current function correctly even if the target line resolves to multiple addresses. (e.g. debugging a templated function)

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@190572 91177308-0d34-0410-b5e6-96231b3b80d8