Adds an option to resolve a symbol from an address that can be used to build out the symbol table as addresses are used, and implements the mechanism for ELF to add stripped symbols from eh_frame.
Uses this mechanism to allow disassembly for addresses corresponding
to stripped symbols for ELF, and provide hooks to implement this for
PE COFF.

Also removes eSymbolContextTailCall in favor of an option for
ResolveSymbolContextForAddress for consistency with the documentation
for eSymbolContextEverything.  Essentially, this is just an option for
interpreting the so_addr.
                  


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@191307 91177308-0d34-0410-b5e6-96231b3b80d8