<rdar://problem/15191078>
Fixed Module::ResolveSymbolContextForAddress() to be able to also look in the SymbolVendor's SymbolFile's ObjectFile for a more meaningful symbol when a symbol lookup finds a synthetic symbol from the main object file. This will help lookups on MacOSX as the main executable might be stripped, but the dSYM file always has a full symbol table.




git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@192510 91177308-0d34-0410-b5e6-96231b3b80d8