LLDB can crash if given DWARF debug info for a class that has a base class which isn't a complete definition.
<rdar://problem/15594781>

We need to not crash at any cost. We currently detect if any base classes are forward declarations, emit an error string that directs the use to file a compiler bug, and continues by completing the class with no contents. This avoids a clang crash that would usually follow when we call setBase().



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@197108 91177308-0d34-0410-b5e6-96231b3b80d8