TestPtrRef2Typedef.py change - take 2.
Our build is still getting this test to fail with the presence of const in
'(Foo &const) y = 0x' and '(Foo &&const) z = 0x'.  This change simply changes
the match substrings to ignore the portion after the final '&' and before the
closing ')'.  This should cover the presence (or absence) of the const qualifier.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@206762 91177308-0d34-0410-b5e6-96231b3b80d8