Added gdb remote test for one valid qRegisterInfo result.
Checks that at at least qRegisterInfo0 responds with a valid
register info reply packet.  The packet is parsed and validates
that all keys come from the documented set of valid keys.  It
then validates that a minimum set of expected keys
are present in the returned packet.

This test is set to pass on debugserver and fail on llgs TOT.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@209109 91177308-0d34-0410-b5e6-96231b3b80d8