[cmake] Switch python install to use an 'install(DIRECTORY...)' cmake command instead of a script.
In addition to cleaning things up, this allows more easy access to the
variables. In the old version, it tried to pass variables as -D flags to
cmake, but this didn't actually work. CMake drops all of those arguments
on the floor (try passing garbage through them) and just picks up the
limited subset of pre-defined macros. So, for example, this fixes the
build with LLVM_LIBDIR_SUFFIX=64 which is how I ended up here. =]

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@211028 91177308-0d34-0410-b5e6-96231b3b80d8