Remove extra newline from log Printf
Clean up this one specifically, as it has the effect of double-spacing
the list of thread stop reasons, and substantially bloats the log file
when opening a core with hundreds of threads.

There are other cases of extra newlines.  Some of them do increase
readability, so avoid a general sweep for now.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@211655 91177308-0d34-0410-b5e6-96231b3b80d8