llgs: implement --setsid.
The --setsid (-S) option changes the session id for the lldb-gdbserver process.
This is used by tools such as lldb-platform and allows the user to prevent
llgs from being in the same session as a calling terminal session.
This will prevents terminal group control signals from affecting
lldb-gdbserver.

See also:
https://github.com/tfiala/lldb/issues/38


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@212873 91177308-0d34-0410-b5e6-96231b3b80d8