DataFormatters: add missing function implementations
This adds a definition for the TypeValidatorImpl_CXX destructor.  Because the
destructor is first virtual method, and declared out-of-line, it also serves as
the key function.  Since no definition was present, no virtual table for
TypeValidatorImpl_CXX was emitted, which results in link failures due to
references to undefined symbols.

Also add a definition for a TypeValidatorImpl contructor which was declared
out-of-line and referenced in a constructor for TypeValidatorImpl_CXX.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@217375 91177308-0d34-0410-b5e6-96231b3b80d8