remove a couple of default cases from switches
This cleans up a couple of warnings [-Wcovered-switch-default] from the build by
removing the default case from a couple of switches which are fully covered.
This is generally better as it will help identify when a new item is added to
the enumeration but the use sites are not updated.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@217376 91177308-0d34-0410-b5e6-96231b3b80d8