ELF: store the thumbness of a function in symbol flags.
This allows us to fixup the address of the symbol as soon as we parse it
so that lldb is not confused thinking there are two different symbols in
the binary (one with the thumb bit, one without). Also, differentiating
between THUMB and ARM symbols allows the debugger to place the right
type of breakpoint.

Change by Stephane Sezer.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@217841 91177308-0d34-0410-b5e6-96231b3b80d8