Added a new kind of test case: the "inline" test case.  This test case style attempts to shed all of the boilerplate that is required for test cases, and let 80% of test cases use a much terser syntax.
Inline testcases have much simplified python files
(the corresponding .py file should contain two
lines of code) and require no Makefile, because the
Makefile is generated automatically.  Breakpoints
are set automatically and the indicated breakpoint
actions (specified after a magic //% comment) are
executed when the breakpoint is hit.

All other testcases are unaffected.

One thing I'm not really happy with yet is the way
multiple actions for the same line are specified.
I'm going to use lang/c/struct_types as a guinea
pig to develop this further.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@219984 91177308-0d34-0410-b5e6-96231b3b80d8