Skip dsym test on !Darwin
r219978 fixed this test to work on Darwin, and removed the expected
failure decorator, but it then started running (and failing) on FreeBSD.

I'd think @dsym_test should skip the test on all non-Darwin operating
systems. It seems not to be the case, so for now skip it the same way as
done for other @dsym_test tests.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@220219 91177308-0d34-0410-b5e6-96231b3b80d8