Fix some compiler warnings, one of which was a legit bug.
MSVC warns that not all control paths return a value when a switch
doesn't have a default case handler.  Changed explicit value checks
to a default check.

Also, it caught a case where bitwise AND was being used instead of
logical AND.  I'm not sure what this fixes, but presumably it is
not covered by any kind of test case.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@221636 91177308-0d34-0410-b5e6-96231b3b80d8