Sketch out the armv7 and arm64 core file writing support in ObjectFileMachO.  It's close but we seem to be missing some of the memory region segments - not exactly sure how that's happening.  The register context writing into the LC_THREAD load commands is working correctly though.
Slightly reordered the arm64 definitions in ArchSpec.cpp so
when we look for an arm64 core file definiton we're getting
a cpu subtype of CPU_ANY which we can't put in the mach
header of a core file.  Make the first definition we find by
linear search have the currently correct '1' cpu subtype.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@221743 91177308-0d34-0410-b5e6-96231b3b80d8