[ProcessWindows] Create a TargetThreadWindows class.
This creates a TargetThreadWindows class and updates the thread
list of the Process with the main thread.  Additionally, we
fill out a few more overrides of Process base class methods.  We
do not yet update the thread list as threads are created and/or
destroyed, and we do not yet propagate stop reasons to threads as
their states change.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@222148 91177308-0d34-0410-b5e6-96231b3b80d8