Read the LSDA and Personality Routine function address out of the eh_frame data.  These two pieces of information are used in the process of exception handler unwinding on SysV ABI systems.
This patch reads the data from the eh_frame section 
(DWARFCallFrameInfo.cpp), allows for it to be saved & read out
of a given UnwindPlan (UnwindPlan.h, UnwindPlan.cpp) - as well
as printing the information in the UnwindPlan::Dump method - and
adds methods to the FuncUnwinders object so that higher levels
can query if a given function has an LSDA / personality routine
defined.

It's only lightly tested, but seems to be working correctly as long
as your have this information in eh_frame.  Does not address getting
this information from compact unwind yet on Darwin systems.

<rdar://problem/18742797> 



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@222214 91177308-0d34-0410-b5e6-96231b3b80d8