Add a new 'eRegisterInLiveRegisterContext' RegisterLocation to track a register value that is live in the stack frame 0 register context.
Fixes a problem where retrieving a register value on stack frame #n
would involved O(n!) stack frame checks.  This could be very slow on
a deep stack when retrieving register values that had not been
modified/saved by any of the stack frames.  Not common, but annoying
when it was hit.

<rdar://problem/19010211> 


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@223843 91177308-0d34-0410-b5e6-96231b3b80d8