Removed the assertion that we can find any named Objective-C type in the runtime.  This is not actually true, it's entirely possible to say
@class DoesntExist;

@interface DoesExist {
  DoesntExist *whyyyyy;
}
@end

and this code will not only compile but also run.  So
this assertion will fire in situations users might
encounter.

I left the assertion enabled in debug mode, because we
could still catch a case we're not aware of (i.e., a
class that we *ought* to have found but where somehow
we mis-parsed the name).

<rdar://problem/19151914>


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@224038 91177308-0d34-0410-b5e6-96231b3b80d8