Fix about 20 tests on Windows.
Passing the argument string from dosep to dotest was failing on
Windows due to the fact that Windows uses \ for its path separator.
As a result, shlex.split() was treating it as an escape character.
This fixes the issue by telling shlex.split() to not use posix mode
when running on Windows.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@225195 91177308-0d34-0410-b5e6-96231b3b80d8