Add support for SBProcess::PutSTDIN to remote processes
Processes running on a remote target can already send $O messages
to send stdout but there is no way to send stdin to a remote
inferior.

This allows processes using the API to pump stdin into a remote
inferior process.

It fixes a hang in TestProcessIO.py when running against a remote
target.



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@228419 91177308-0d34-0410-b5e6-96231b3b80d8