Make a more complete fix for always supplying an execution context when getting byte sizes from types.
There was a test in the test suite that was triggering the backtrace logging output that requested that the client pass an execution context. Sometimes we need the process for Objective C types because our static notion of the type might not align with the reality when being run in a live runtime.

Switched from an "ExecutionContext *" to an "ExecutionContextScope *" for greater ease of use.



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@228892 91177308-0d34-0410-b5e6-96231b3b80d8