Resubmit "[CMake] Change lldbAPI to be a CMake OBJECT library."
This resubmits r230380.  The primary cause of the failure was
actually just a warning, which we can disable at the CMake level
in a followup patch on the LLVM side.  The other thing which was
actually an error on the bot should be able to be fixed with
a clean.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@230389 91177308-0d34-0410-b5e6-96231b3b80d8