Add filepath to qModuleInfo packet
The file path is currently required on android because the executables
only contain the name of the system libraries without their path. This
CL add an extra field to the qModuleInfo packet to return the full path
of a modul and add logic to locate a shared module on android.

Differential revision: http://reviews.llvm.org/D8221

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@232156 91177308-0d34-0410-b5e6-96231b3b80d8