Rework the gtest directory structure.
This makes the directory structure mirror the canonical LLVM
directory structure for a gtest suite.

Additionally, this patch deletes the xcode project.  Nobody
is currently depending on this, and it would be better to have
gtest unit tests be hand-maintained in the Xcode workspace
rather than using this python test runner.  Patches to that
effect will be submitted as followups.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@232211 91177308-0d34-0410-b5e6-96231b3b80d8