Fix FreeBSD crash after r233837
The FreeBSD debug register access is a little usual, but in any case
different from Linux. As it stands it's not possible to share an
implementation of DR_OFFSET, so revert that part of r233837 and provide
a separate FreeBSD and Linux implementation.

We'll still want a better fix, but this should restore basic
functionality (and the buildbot).


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@234048 91177308-0d34-0410-b5e6-96231b3b80d8