Fix test expectation in TestNoreturnUnwind
The test case lookinhg for the abort function in the stack trace.
Previously it lookd for a function which ends with "abort" but on some
system there are multiple such functions (e.g.: on android abort calls
__libc_android_abort) what made the test fail. This CL change the
behaviour to look for the abort function based on a fix list of names.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@235584 91177308-0d34-0410-b5e6-96231b3b80d8