Remove the NullLog class introduced in r236174.
Based on list discussions, a different approach is desired for
reducing the visual impact of logging statements on the
readability of the code.  Another mechanism will be added in
a followup patch, but for now, since NullLog is unreferenced,
this patch just removes it.

This patch does *not* remove the other half of r236174, which was
to delete some dead code surrounding logging flags.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@236259 91177308-0d34-0410-b5e6-96231b3b80d8