Cleanup the code a bit to make it more readable.
Add some if/then to avoid calling a function to get dynamic/synthetic types if we know we aren't going to need to call it.

Avoid calling a function that returns a shared pointer twice: once for testing it and once for assigning it (even though that shared pointer is cached inside the value object), it just makes the code a bit clearer.



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@240299 91177308-0d34-0410-b5e6-96231b3b80d8