[MainLoop] Fix assertion failure
Upon connection termination the waitable handle of an IOObject gets reset to an invalid handle.
This caused a problem since we used the object->GetWaitableHandle as a key to the set of
registered events. The fix is to use something more immutable as a key: we make a copy of the
original waitable handle, instead of holding onto the IOObject.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@242515 91177308-0d34-0410-b5e6-96231b3b80d8