Report original thread ID for FreeBSD core files
On FreeBSD the tid is (somewhat unintuitively) found in the pr_pid
field of the NT_PRSTATUS note. Collect it when parsing the note and
store it in the thread data.

For Linux I've left the original behaviour of using sequential TIDs
(0, 1, 2...) as I don't yet have code to obtain it.

Differential Revision:	http://reviews.llvm.org/D11652


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@243748 91177308-0d34-0410-b5e6-96231b3b80d8