Fix TestArrayTypes on Windows.
Whether or not frames print their tid in hex or decimal is apparently
hardcoded to depend on the operating system.  For now a comment was
added that this should be changed to a more sane check (for example
a setting), and the OS check is updated to do the right thing for
Windows.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@245371 91177308-0d34-0410-b5e6-96231b3b80d8