Fixup XFAIL marking on TestConstVariables for clang version
Linux/clang on clang 3.6+ is passing this test.  As OS X was
in a similar position, I'm guessing this was fixed by a recent
compiler.  Might work on 3.5/3.4, but it is definitely working
on clang 3.6.  I didn't break out FreeBSD since this appears to
be related to a compiler fix.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@247633 91177308-0d34-0410-b5e6-96231b3b80d8