Fix prologue end handling when code compiled by gcc
GCC don't use the is_prologue_end flag to mark the first instruction
after the prologue. Instead of it it is issuing a line table entry for
the first instruction of the prologue and one for the first instruction
after the prologue. If the size of the prologue is 0 instruction then
the 2 line entry will have the same file address.

We remove these duplicates entries as they are violating the dwarf spec
and can cause confusion in the debugger. To prevent the lost of
information about the end of prologue we should set the prologue end
flag for the line entries what are representing more then 1 entry.

Differential revision: http://reviews.llvm.org/D12757

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@247788 91177308-0d34-0410-b5e6-96231b3b80d8