test events: added optional value type to extra event key/val pairs
The test events had worker indexes coming across as strings.  I
want them to be ints.  worker_index now comes across as an int in
the dicationary.

The optional type can be specified with:
--event-add-entries key=val[:type][,key2=val2[:type2]...]
The type piece may be 'int' at this time.  That is all.  Otherwise
it will be a string.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@248066 91177308-0d34-0410-b5e6-96231b3b80d8