[LLDB][MIPS] Fix gp register value for o32 applications on 64-bit target
GP registers for o32 applications were always giving zero value because SetType() on the RegisterValue was causing the accessor functions to pickup the value from m_scalar of RegisterValue which is zero.
In this patch byte size and byte order of register value is set at the time of setting the value of the register.



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@249020 91177308-0d34-0410-b5e6-96231b3b80d8