Make dwarf parsing multi-threaded
Loading the debug info from a large application is the slowest task
LLDB do. This CL makes most of the dwarf parsing code multi-threaded.

As a result the speed of "attach; backtrace; exit;" when the inferior
is an LLDB with full debug info increased by a factor of 2 (on my machine).

Differential revision: http://reviews.llvm.org/D13662

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@250821 91177308-0d34-0410-b5e6-96231b3b80d8