Remove use of octal literals.
Python 3 has a different syntax for octal literals than Python 2
and they are incompatible with each other.  Six doesn't provide
a transparent wrapper around this, so the most sane thing to do
is to not use octal literals.  If you need an octal literal,
use a decimal literal and if it's not obvious what the value is,
provide the value in octal as a comment.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@251328 91177308-0d34-0410-b5e6-96231b3b80d8