Move lldb/test to lldb/packages/Python/lldbsuite/test.
This is the conclusion of an effort to get LLDB's Python code
structured into a bona-fide Python package.  This has a number
of benefits, but most notably the ability to more easily share
Python code between different but related pieces of LLDB's Python
infrastructure (for example, `scripts` can now share code with
`test`).

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@251532 91177308-0d34-0410-b5e6-96231b3b80d8