Revert r251578.  I was in the middle of editing the commit message and decided to not commit the change, but accidentally committed it anyway.
This was based on a static analysis complaint; it thought there
was a code path where AdbClient::PushFile would call AdbClient::ReadSyncHeader
and AdbClient::ReadSyncHeader wouldn't set data_len.  But in that
case, the Error object returned will be Fail and we won't use
the data_len or response_id contents.



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@251580 91177308-0d34-0410-b5e6-96231b3b80d8