Fix new summary to include exceptional exit count in determining exit value
The main dotest.py should exit with a system return code of 1 on any
issue.  This change fixes a place where I omitted counting the
exceptional exit value to determine if we should return 1 when using the
new summary results.

This change also puts a banner around the Issue Details section that comes
before the Test Result Summary.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@255138 91177308-0d34-0410-b5e6-96231b3b80d8