test infra: force rerun to use parallel runner
We've now seen the rerun test phase hang in a few
scenarios.  Eliminate the serial test runner (which
is not exercised nearly as much as the others), by
using a multi-worker test runner strategy with a single
worker.  This should rule out whether this is related
to the serial test runner strategy.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@255880 91177308-0d34-0410-b5e6-96231b3b80d8