Fiddling with Pavel's fix for getting the breakpoint right when there's debug info for sleep.
I prefer to use "-p" over using line_number and then setting by line because it's makes it possible
to see what the breakpoint is at the site where you make the breakpoint.  So I switched
it back to -p but specified the source file as well, which is an "all within lldb" way of doing
what Pavel's fix did.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@257235 91177308-0d34-0410-b5e6-96231b3b80d8