Silence an incorrect dwarf parsing warning
We have a check what warns if the offset of a class member is greater
then or equal to the size of the class. The warning is valid in most
case but it is invalid when the last data member is a 0 size array
because in this case the member offset can be equal to the class size
(subject to alignment limitations).

This CL fixis LLDB to not print out a warning in this special case.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@257603 91177308-0d34-0410-b5e6-96231b3b80d8