Enable the use of modules in the expression parser by default.
If your program refers to modules (as indicated in DWARF) we will now try to
load these modules and give you access to their types in expressions.  This used
to be gated by a setting ("settings set target.auto-import-clang-modules true")
but that setting defaulted to false.  Now it defaults to true -- but you can
disable it by toggling the setting to false.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@257812 91177308-0d34-0410-b5e6-96231b3b80d8