When generating result variables, look through the "$__lldb_context" typedef.
Currently, if you return "self" from an expression, you'll get its type showing
up as "$__lldb_context", which is an artifact of the way the expression is set
up for parsing.  This is never what you actually want to see, so now the
expression parser removes the sugar for that particular typedef and you get the
real type back.