Fix compiler lookup when specified without path
r259433 introduced a regression, where if a compiler is specified without a path (e.g., CC=clang,
relying on the fact that clang is in $PATH), then the test suite would fail (at the compiler
version detection step) because realpath would interpret this as a path relative to cwd). The fix
is to perform the $PATH expansion (via `which`) before the realpath step.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@259484 91177308-0d34-0410-b5e6-96231b3b80d8