Don't return a tuple from the skip test function.
Previously we were returning a tuple of (bool, skip_reason) from
the tuple function.  This makes for some awkward code, especially
since a value of True for the first argument implies that the
second argument is None, and a value of False implies that the
second argument is not None.  So it was basically redundant, and
with this patch we simply return the skip reason or None directly.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@259590 91177308-0d34-0410-b5e6-96231b3b80d8