normalize test file extension for test filenames (#20)
It looks like something changed so that a test's filename can
report as the .pyc rather than the .py file in some flows.  This
causes confusion, particularly in cases where a test runner
queue has to synthesize a test event, uses the .py extension that
we always used.  Then the rerun result event has a different
key than the timeout test run, and the timeout test run does
not get overwritten.

This change introduces a normalization step that convers all
test filenames ending in .pyc to end in .py.

Fixes:
https://bugs.swift.org/browse/SR-1220