Expressions can run without a process.
Code was added in ClangExpressionParser::ClangExpressionParser that was calling through
the process w/o checking that it was good.  Also, we were pretending that we could do something
reasonable if we had no target, but that's actually not true, so I check for a target at the
beginning of the constructor and don't make a compiler in that case.

<rdar://problem/25841198>


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@266944 91177308-0d34-0410-b5e6-96231b3b80d8