Add a way for an ObjectFile to indicate that assembly emulation should not be used for this module -- for use when an ObjectFile knows that it does not have meaningful or accurate function start addresses.
More commonly, it is not clear that function start addresses are
missing in a module.  There are certain cases on Mac OS X where we
can tell that a Mach-O binary has been stripped of this essential
information, and the unwinder can end up emulating many megabytes
of instructions for a single "function" in the binary.

When a Mach-O binary is missing both an LC_FUNCTION_STARTS load 
command (very unusual) and an eh_frame section, then we will assume 
it has also been stripped of symbols and that instruction emulation
will not be useful on this module.

<rdar://problem/25988067> 


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@268475 91177308-0d34-0410-b5e6-96231b3b80d8