Don't let two threads call Debugger::Clear simultaneously.
We don't want a mutex in debugger as it will cause A/B locking issues with the lldb_private::Target's mutex, but we do need to stop two threads from doing Debugger::Clear at the same time. We have seen issues with this with the C++ global destructor chain where the global debugger list is being destroyed and the Debugger::~Debugger() is calling it while another thread was in the middle of running that function.

<rdar://problem/26098913>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@268563 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 0b8ef1574e88d0cd62cca3d69ea53091fa42f20c)