Fix logging in Listener.cpp
Clear() log message was claiming it was the destructor, which had me very confused when looking
at the log messages. Fix the message, and add a log message to the real destructor.

Also noticed that the destructor was needlessly locking the broadcaster mutex (as Clear was
locking it again anyway), so remove that as well.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@269058 91177308-0d34-0410-b5e6-96231b3b80d8