Add support for arm64 compact unwind tables, used on darwin arm64 systems (ios, tvos, watchos).  It's a simple format to use now that I have i386/x86_64 supported already.
The unwind instructions are only valid at call sites -- that is,
when lldb is unwinding a frame in the middle of the stack.  It
cannot be used for the currently executing frame; it has no information
about prologues/epilogues/etc.

<rdar://problem/12062336>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@270658 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 94ef55a6048ba44fa00cf5af21e357e002efe671)