Add more verification on consectutive bitfields otherwise clang will assert.
We need to verify that consecutive bitfields have higher offsets and don't overlap. The issues was found by running a broken version of recent clangs where the bitfield offsets were being emitted incorrectly. To guard against this we now verify and toss out any invalid bitfields and print a message that indicates to file a bug against the compiler.

<rdar://problem/25737621>



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@271343 91177308-0d34-0410-b5e6-96231b3b80d8