[tsan] Prefer mangled name looking up variable declaration for racy address
For Thread Sanitizer reports, LLDB tries to find a global variable declaration
corresponding to the racy address in order to provide a filename and line
number. This commit changes the lookup of the variable to use the mangled
name for lookup and fall back to the demangled version if unavailable. This
is needed to report locations of races on Swift global variables.

I've also added a test to make sure we look up C++ globals correctly.

rdar://problem/26459401

Differential Revision: http://reviews.llvm.org/D20760

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@271433 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 0adc80f6139ec87e9e787a325c376c04314ac8a9)