(Minor tweak) Make RegisterContextWindows_x86/x64::GetRegisterInfoAtIndex return NULL for an invalid register.
The unwind logic asks for the "return address register" which doesn't exist
on x86/x86_64, returns -1 and calls this with -1 as a parameter, ends up 
out of scope of the array bounds for g_register_infos and later SIGSEGVs 
on accessing. This now matches the other GetRegisterInfoAtIndex for
other platforms.



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@271876 91177308-0d34-0410-b5e6-96231b3b80d8