Add support for using armv7 compact unwind information as an asynchronous unwind plan source.
Two small fixes to the compact unwind dumper tool for
armv7 encodings.

A change to DWARFCallFrameInfo to strip the 0th bit on
addresses in eh_frame sections when armv7.  In the 
clang generated examples I have, the 0th bit is set for
thumb functions and that's causing the unwinder to pick
the wrong function for eh_frame info.




git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@271970 91177308-0d34-0410-b5e6-96231b3b80d8