Enable some tests on linux
This enables a couple of tests which have been shown to run reliably on the
linux x86 buildbot. If you see a failure after this commit, feel free to add
the xfail back, but please make it as specific as possible (i.e., try to make
it not cover i386/x86_64 with clang-3.5, clang-3.9 or gcc-4.9).

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@272326 91177308-0d34-0410-b5e6-96231b3b80d8