Set TERM env var for Xcode gtests
Not sure what changed, but something outside our code
is failing one of the EditLine gtests on OS X CI (and
locally) before the gtest ever gets to run.  This fails
the first EditLine gtest.

This change exports the TERM as "vt100" before running
the lldb-gtest binary, fixing the issue.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@272844 91177308-0d34-0410-b5e6-96231b3b80d8