Allows "experimental" settings that will either route to their containing settings or raise no error if not found.
From time to time it is useful to add some setting to work around or enable
a transitory feature.  We've been reluctant to remove them later because then
we will break folks .lldbinit files.  With this change you can add an "experimental"
node to the settings.  If you later decide you want to keep the option, just move
it to the level that contained the "experimental" setting and it will still be
found.  Or just remove it - setting it will then silently fail and won't halt
the .lldbinit file execution.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@274593 91177308-0d34-0410-b5e6-96231b3b80d8