If the main executable is called 'XCTest,' configure Swift using the test bundle
When running unit tests in Xcode, the main unit test executable does not have
any Swift debugging information.  Instead, that information is in the test
bundle.  We need to configure the expression parser using this executable and
ignore all other sources of debug information.

This patch adds this support and also a testcase.

<rdar://problem/26291135>