force std::regex_replace() to use std::string version
This was failing on Ubuntu 14.04 x86_64 (although not on
15.10 or 16.04) with the version using const char*.
The spec indicates that should be present in C++11, but
the compiler was not presenting the variant on 14.04.
I switched over to using the std::string variant of
std::regex_replace() to try to work around the 14.04
compiler/c++ stdlib issue.