Add support for an additional dictionary in the per-arch plists that may be embedded in the Contents/Resources subdir of a dSYM bundle.  These allow for the specification of a build-time path to debug-time path remapping for source files.  Files may be built in /BuildDirectory/sources/project-100 but when the debugger is run, they're actually found via ~sources/project-100 - this plist allows for that remapping through the DBGBuildSourcePath and DBGSourcePath keys.
This patch adds support for a new DBGSourcePathRemapping
dictionary in the plist where the keys are the build-time paths
and the values are the debug-time paths that they should be
remapped to.  There are instances were we have multiple possible
build-time paths that need to be included, so the dictionary was
required.

<rdar://problem/26725174>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@276729 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit a17cfde236215d95b4adffc463f7c0282501e680)