Ignore uninitialized generic variables in expressions, fixing a spurious error.
The Swift expression parser knows to ignore variables that can't be read.
However, variables of generic type, when uninitialized, appear to be readable
until you try to get their fully-resolved type.  This was causing the expression
parser to erroneously expose them to expressions even though they are unusable.
This patch fixes that and adds a testcase.

<rdar://problem/27669166>