Rewrite the GetFileInSDK methods in PlatformRemoteiOS, PlatformRemoteAppleWatch, PlatformRemoteAppleTV and remove the GetFileInSDKRoot method from those classes.
The rewrite uses the more modern FileSpec etc API to simplify,
and handles the case where an SDK Root is given to lldb with
the "/Symbols" directory name already appended.  The new version
will try appending "/Symbols" and "/Symbols.Internal" to the 
sdk root directories, and will also try appending nothing to
the sdk root directory in case it's handed such an sdkroot.

<rdar://problem/28000054> 


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@279688 91177308-0d34-0410-b5e6-96231b3b80d8