Convert GetNameColonValue to return StringRefs.
StringExtractor::GetNameColonValue() looks for a substring of the
form "<name>:<value>" and returns <name> and <value> to the caller.
This results in two unnecessary string copies, since the name and
value are not translated in any way and simply returned as-is.

By converting this to return StringRefs we can get rid of hundreds
of string copies.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@280000 91177308-0d34-0410-b5e6-96231b3b80d8