Disable using socketpair sockets on Apple hardware
r278524 brought in support for socketpair() usage within debugserver
and a new "--fd" file descriptor arg accepted on the command line.
This is fine for when lldb and debugserver are in sync.  We have
several CI installations that are required to use an Apple signed
debugserver, and thus lag behind the lldb client that is running
tests.  This causes a catastrophic failure in the test suite when
lldb fails to launch debugserver (using args that aren't supported
on the Xcode-provided, Apple-signed debugserver).

This change comments out the #define that enables the new behavior.
We can back this out once CI are building with an Apple-signed
debugserver that contains the "--fd" support.