Fix unittest compilation on windows
After the reformat, the unittests do not compile due to missing due to redefinition errors
between PosixApi.h and ucrt/direct.h. This is a bit of a shot in the dark, as I have not tested
it on windows, but I am restoring the original include order, so it should hopefully fix it.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@280793 91177308-0d34-0410-b5e6-96231b3b80d8