Fix test breakage in r280919
It turns out that self.dbg.GetSelectedPlatform().GetTriple() is not a good way
to get the triple of the process, as it returns the incorrect triple in case of a
32-bit process running on a 64-bit platform.

Instead, go the long way round and ask the stub for the process triple. This
fixes the test for i386.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@280922 91177308-0d34-0410-b5e6-96231b3b80d8