Reduce the number of simultaneous debug sessions to 10 and remove the expectedFlakeyDarwin annotation.
I've been running this test in isolation on my macOS Sierra system
and haven't seen a failure in 20-30 runs.  The number of simultaneous
debug sessions that it spins up could be a problem when the testbots
are running under load, so I'm reducing this from 20 simultaneous
debug sessions to see if we can get enough stability to leave this
enabled.



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@281291 91177308-0d34-0410-b5e6-96231b3b80d8