added LLDB_PYTHON_TESTSUITE_ARCH Xcode variable
This Xcode build variable defaults to x86_64.  It can be set to i386
to cause the lldb-python-test-suite target run the tests in the
specified architecture.

This flag is being added for the zorg build script so that Green Dragon
can run the test suite against both x86_64 and i386 macOS targets.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@281639 91177308-0d34-0410-b5e6-96231b3b80d8