Fix failing regex tests.
r282079 converted the regular expression interface to accept
and return StringRefs instead of char pointers.  In one case
a null pointer check was converted to an empty string check,
but this was an incorrect conversion because an empty string
is a valid regular expression.  Removing this check should
fix the test failures.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@282090 91177308-0d34-0410-b5e6-96231b3b80d8