fix DarwinLog no-info/no-debug case
This started failing recently:
TestDarwinLogSourceDebug.py

It looks like the behavior of specifying the OS_ACTIVITY_MODE
env var with no value used to work for no-info/no-debug content.
That doesn't appear to be the case now.  Switch to specifying
the proper value ('default') when no info-level and no debug-level
content is expected.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@282172 91177308-0d34-0410-b5e6-96231b3b80d8