[CMake] Fix LLDB in-tree build
This is a bunch of small fixes that enable swift-lldb to build in-tree. This change is NFC for build-script builds. With this change in-tree builds do not depend on all Clang and Swift static archives, instead the build only depends on a minimal set. Additionally the in-tree builds add many required dependencies on clang and swift targets that are not needed in the build-script builds.