[CMake] Populate the build directory's framework
This ensures that the Resources and clang headers are properly symlinked in LLDB's framework. This should fix the modules-related tests when building on Darwin with CMake if you are building a framework.

I have another fix coming which gets them working on Darwin if you're building liblldb instead of a framework.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@285651 91177308-0d34-0410-b5e6-96231b3b80d8