Switch SBWatchpoint::SetEnabled over to using Process::{Enable,Disable}Watchpoint.
We don't have a good story for what happens to watchpoints when you don't
have a process, or if your process exits.  Clearing that up will instruct 
how to fix this for real.

Also added a test to make sure disable->enable works as well.
This resolves llvm.org/pr30789.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@285742 91177308-0d34-0410-b5e6-96231b3b80d8