Fixed a crasher that has been borking out heap for a long time.
ThreadList had an assignment operator that didn't lock the "rhs" thread list object. This means a thread list can be mutated while it is being copied.

The copy constructor calls the assignment operator as well. So this fixes the unsafe threaded access to ThreadList which we believe is responsible for a lot of crashes.

<rdar://problem/28075793>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@289100 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 1b0a6070fd2efd2a5b06f9833ed558ebe8f09b53)