Fix-up TestLinuxCore for r290874
r290874 enabled the s390x test, which caused the rest of the tests to start
misbehaving. This is because this test switches the selected platform and the
change persists.

This fixes it by explicitly resetting the platform in a similar way to the gcore
tests do. Potentially we should consider re-setting the platform globally
between each test run to better protect tests from each other.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@290890 91177308-0d34-0410-b5e6-96231b3b80d8