Made repl_swift build with the C compiler under Xcode.
In some environments the Swift compiler is unusable at the time LLDB is
being built, and the Swift compiler LLDB builds is also not usable.  In this
case making repl_swift a simple C binary is preferable.

It still needs to load the core Swift library so that LLDB can find the
necessary symbols, so I've ensured that the Swift library is linked in and
actually loaded by the dynamic loader.