[lldb] Add support for "external" reports in ThreadSanitizer LLDB plugin
TSan now has the ability to report races on "external" object, i.e. any library class/object that has read-shared write-exclusive threading semantics. The detection and reporting work almost out of the box, but TSan can now provide the type of the object (as a string). This patch implements this into LLDB.

Differential Revision: https://reviews.llvm.org/D30024



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@295342 91177308-0d34-0410-b5e6-96231b3b80d8