NPL: Fix single step workaround
While refactoring the code in r293046 I made a very basic error -
relying on destructor side-effects of a copyable object. Fix that and
make the object non-copyable.

This fixes the tests on the platforms that need this workaround, but
unfortunately we don't have a way to make a more platform-agnostic test
right now.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@295345 91177308-0d34-0410-b5e6-96231b3b80d8