Changed builld-llvm.py to use .json files
LLDB has many branches in a variety of repositories.
The build-script.py file is subtly different for each set.
This is unnecessary and causes merge headaches.

This patch makes build-llvm.py consult a directory full 
of .json files, each one of which matches a particular
branch using a regular expression.

Differential revision: https://reviews.llvm.org/D30275


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@295897 91177308-0d34-0410-b5e6-96231b3b80d8