[debugserver] NFC. Cleanup DNBArchImpl*::GetFPUState()
This patch consolidates the DEBUG_FPU_REGS code for i386 and x86_64 to take advantage of the fact that the non-AVX members of the avx register state structure overlap with the standard fpu register state structure.

This reduces the amount of code required to set debug values into the register state structures because the register state structures are stored in a union.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@297688 91177308-0d34-0410-b5e6-96231b3b80d8