Fix flakyness in TestCompletion
One of the file name templates was occasionally generating the name
"fooa***", which conflicted with the one of the tests expectation that
there is only one item beginning with "fooa".

There doesn't seem to be a good reason for using random file templates
here, so just switch to a fixed set of files to increase
reproducibility.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@297743 91177308-0d34-0410-b5e6-96231b3b80d8