Use two LC_NOTE load commands for identifying the main binary that lldb should use when given a corefile.
This uses an LC_NOTE "main bin spec" or an LC_NOTE "kern ver str"
if they are present in a Mach-O core file.

Core files may have multiple different binaries -- different kernels,
or a mix of user process and kernel binaries -- and it can be
difficult for lldb to detect the correct one to use simply by looking
at the pages of memory.  These two new LC_NOTE load commands allow
for the correct binary to be recorded unambiguously.

<rdar://problem/20878266> 


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@300138 91177308-0d34-0410-b5e6-96231b3b80d8