Disable a test in TestReturnValue on arm64 linux
as described in pr33042, we cannot reliably retrieve the return value on
arm64 in cases it is returned via x8 pointer. I tried to do this as
surgically as possible and disabled it only on targets I know to be
affected, as the code is still useful, even though it can only work on
best-effort basis.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@303076 91177308-0d34-0410-b5e6-96231b3b80d8