[TypeSystem] Fix inspection of Objective-C object types
ptr_refs exposed a problem in ClangASTContext's implementation: it
uses an accessor to downcast a QualType to an
ObjCObjectPointerType, but the accessor is not fully general.
getAs() is the safer way to go.

I've added a test case that uses ptr_refs in a way that would
crash before the fix.

<rdar://problem/31363513>


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@303110 91177308-0d34-0410-b5e6-96231b3b80d8