Simplify the task of stepping out from code with source.
We were stepping through (and using step IN) to get through code
that steps out from somewhere we don't want to stop.  This simplifies
that by stepping out when possible, or stepping over rather than in.

Also, if a StepOut plan ends up stepping past the place it started,
be sure to update the step out to frame to the new frame you are
stepping out to.

<rdar://problem/31515444>