Fix GreenDragon bots
This commit removes a very old deprecated API that was causing compile failures for LLDB on Darwin. Since the comment says we only needed to keep the old API around for a few Xcode builds, and the comment was written 6 years ago... I think this can safely go away.

Failure URL:
http://lab.llvm.org:8080/green/view/LLDB/job/lldb_build_test/29936/console

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@308509 91177308-0d34-0410-b5e6-96231b3b80d8