Fix Linux remote debugging after r313442
On Linux lldb-server sends an OK response to qfThreadInfo if no process
is started yet. I don't know why would LLDB issue a qfThreadInfo packet
before starting a process but creating a fake thread ID in case of an
OK or Error respoinse sounds bad anyway so lets not do it.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@313525 91177308-0d34-0410-b5e6-96231b3b80d8