Split makefile for TestTopLevelExprs
this test was using a single makefile to build two executables. This
setup, although not supported by Makefile.rules, happened to
work in most configurations, except when building with the android ndk
r16.

Here I move the building of the second executable to a separate
makefile, which is the solution other tests use for multiple targets.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@316982 91177308-0d34-0410-b5e6-96231b3b80d8