Log: delimit thread name in log message
The thread name was not followed by a space, which meant it was glued to
the log message. I also align the name as we do that with other log
fields. I align it to 16 chars instead of llvm::max_thread_name(), as
that can be 64 on darwin, which is rather long. If anybody feels
differently about that, we can change it.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@317679 91177308-0d34-0410-b5e6-96231b3b80d8