Fix Swift dictionary padding with enum keys
When you have a dictionary which has an enum as the key such as:

```
enum Foo: String {
  case a
}

let bar = [Foo.a: 1]
```

The stride of the key is 1 byte. This differs from if you have a tuple
with the same key and value as the dictionary such as:

```
let baz = (Foo.a: 1)
```

In order to read the values from a dictionary we create a tuple which
expects 16 bytes of data (in the case the key is an enum and the value
is an int). Previously because the key was only 1 byte, we only filled
it with 9 bytes, and therefore the offsets were wrong leading to us
printing garbage data. Now instead of just using the stride of the key,
we use the stride based on the tuple type we generate for printing these
values. This works since the buffer is initialized to all 0 and
therefore acts as padding for the key.