Add support for building lldb with ASAN enabled (e.g. xcodebuild -configuration Debug -enableAddressSanitizer YES).
Change the lldb xcode project file to build with the macos SDK
instead of the macos internal SDK (there was no need to build
against the internal SDK).  Change lldb-server to link against
libpanel - building with ASAN seems to change some of the
dead-code-stripping so lldb-server suddenly has a dependency
on libpanel.

Change the check-ast-context.py script to not run when building
with ASAN enabled.

Update the build-llvm.py and lldbbuild.py python scripts to
work correctly when building with ASAN.

An update to the swift lldb build-presets is needed for this to
work; I'll be committing that later today.

<rdar://problem/30567418>