Kill struct IOVEC
struct iovec is used as an interface to system (posix) api's. As such,
we shouldn't be using it in os-independent code, and we shouldn't be
defining our own iovec replacements.

Fortunately, its usage was not very widespread, so the removal was very
easy -- I simply moved a couple declarations into os-specific code.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@319536 91177308-0d34-0410-b5e6-96231b3b80d8