Update PlatformDarwin::GetDeveloperDir to handle the two most common cases where the Xcode.app bundle puts lldb - either as a default part of the bundle, or in a toolchain subdirectory, so the platform subclasses can find files relative to this directory.
Dropped support for handling the case where the lldb
framework was in /Library/PrivateFrameworks.  I think
this was intended to handle the case where lldb is installed
in / (outside the Xcode.app bundle) - but in that case, we
can look in the raw directory file paths to find anything.

<rdar://problem/35285622>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@320240 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit b50ada2be61b2d45fc9733384f9f8129969d7c66)