This change updates the deployment target for lldb and debugserver (built with Xcode) from 10.9 to 10.11.  It also enables the use of libcompression in debugserver by default (these API are only present in macOS 10.11 and newer -- 10.11 was released c. Sep 2015).
I don't know if we have people / bots building lldb on older mac
releases; if this turns out to be a problem I will revert the change.
There are some parts of lldb (e.g. debugserer's ability to report
the OS version #) that only work with 10.10 and this changes the
behavior of lldb (whether the older or newer dyld interfaces are
used) so there is some importance to updating the min required
version.



git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@322128 91177308-0d34-0410-b5e6-96231b3b80d8