Simplify some LogTest tests
This removes boilerplate for setting up a log channel and capturing the
output from some of the tests. I do this by moving the setup code into a
test fixture and adding a logAndTakeOutput utility function to log some
string and then retrieve it from the log.

I also use some googlemock goodies to simplify a couple of assertions.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@322653 91177308-0d34-0410-b5e6-96231b3b80d8