Fix logging test on windows
Refactor in r322653 broke the logging test on windows because MSVC uses
a fully qualified name as the value of __FUNCTION__, which is not what
we were expecting. I'm changing the code to use the more portable
__func__, which behaves the same as on clang/gcc.

The standard still says that the value of __func__ is implementation
defined, so if we run into problems with this test again we may just
need to drop it.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@322834 91177308-0d34-0410-b5e6-96231b3b80d8