[lldb] Fix some C++ virtual method call bugs in LLDB expression evaluation by building method override tables for CXXMethodDecls in DWARFASTParserClang::CompleteTypeFromDWARF.
C++ virtual method calls in LLDB expressions may fail if the override table for
the method being called is not correct as IRGen will produce references to the
wrong (or a missing) vtable entry.

This patch does not fix calls to virtual methods with covariant return types as
it mistakenly treats these as overloads, rather than overrides. This will be
addressed in a future patch.

Review: https://reviews.llvm.org/D41997

Partially fixes <rdar://problem/14205774>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@323163 91177308-0d34-0410-b5e6-96231b3b80d8