Add a comment explaining how the input for GetModuleSpecifications_EarlySectionHeaders was generated
Davide pointed out this would be useful if the file ever needs to be
regenerated (and I certainly agree).

I also replace the test binary with a slightly smaller one -- I intended
to do this in the original commit, but I forgot to add it to the patch
as I was juggling several things at the same time.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@324256 91177308-0d34-0410-b5e6-96231b3b80d8