[SwiftASTContext] Remove now always failing caching of common demangling.
This never kicks in because the mangling prefix changed from `_T` to
`$S`. Also, it's not the debugger job to decide which names are more common,
and it's unsure whether this decision should be static.

I plan to remove this cache altogether, my first set of measurements
show this is not a time sink, and this mechanism exhibits some fragility.

I think a better way of handling this situation is either:
a) Speed up the demangler if it turns out to be slow
b) Have a mechanism in the demangler library that transparently
caches results (and implements some suitable cache evition policy).
This way, the debugger can dynamically communicate what are the set
of symbols more commonly demangled, and the library can decide to
cache them accordingly. This avoids the hardcoding problem and
makes the whole thing more robust.