Shorten socket names in TestPlatformProcessConnect
The test was generating long unix socket names, and the addition of a
new folder in the previous patch pushed it over the limit (although
linux has a fairly generous limit for path names, this does not apply to
unix sockets).

Modify the test to use a shorter name instead.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@325340 91177308-0d34-0410-b5e6-96231b3b80d8