Make gdb-client tests generate binaries in the build tree
These were missed in the great refactor because they were added
concurrently with it. Since we started running tests in a more parallel
fashion they started to be flaky. This should fix it.

Now that we are no longer polluting the source tree, I also delete the
bit of custom cleanup code specific to these tests.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@325495 91177308-0d34-0410-b5e6-96231b3b80d8