[ExpressionParser] Clean up SwiftREPL::CreateInstance
This is a mostly-NFC change which makes SwiftREPL::CreateInstance easier
to read with early returns and references.

The only functional change is a bugfix for the call to
GetScratchTypeSystemForLanguage. If that call fails, we will now
correctly tear down the REPL process and the event handler thread.

(cherry picked from commit 582db9faae909ca55ebcf0038b26f81a80bfed7f)