[SwiftRepl] Improve the CMake build & simplify.
This is a grab bag of improvements for the repl. There are
several issues.
1) The repl needs a dummy target to run. There were two, one
written in C, the other in swift. The latter doesn't really
work and the only reason why nobody noticed is that only the
CMake build used it and there was no test whatsoever. The good
news is that's unneeded and can be removed.
2) The way the dummy executable was built with CMake was wrong
and unnecessarily complicated. This commit rewrites the logic
to be more readable (and commented).
3) This logic expects optimizations to be turned off, but
that was delegated to the build system. Now it's explicit
in the code via attributes, so that we don't have to duplicate
the logic across CMake/XCodebuild.

The logic is still fairly fragile as we dlopen() a library
looking inside @rpath. After speaking with Louis, he told me
that we shouldn't rely on this behaviour (and I tend to agree).
It's more complicated to fix because it involves touching also
the xcodebuild. This is left as a follow-up.