[PATCH] [test] Introduce the skipUnlessSwiftPR decorator
This determines whether or not a test is skipped due to not being marked
`swiftpr` in the wrapper function, i.e after the test method is called,
instead of before.

This fixes strange issues with categories which resulted in the wrong
set of tests being marked `swiftpr` and running accidentally. It also
fixes discrepancies which arise between the single and multi threaded
test runners due to attempts to edit an attribute on a function instance
shared between multiple inline tests.

Thanks to Jonas Devlieghere for debugging the problem here!