[REPL] Rewrite dictionary testing & improve infra to be more reliable.
This now uses lit. It's shorter and hopefully more reliable than
pexpect. As part of this, I also did all the plumbing for testing
the repl in lldb, so what I've done here can be used for other repl
tests. While I'm at it, remove the old test, we don't need it anymore.

<rdar://problem/37407231>