[testsuite] Throw away test/debug_info/apple_types.
This test was only testing that clang produced the correct informations
for __apple accelerated tables. So, it's a clang test. Also, it
doesn't require any debugger intervention, the object file can
be analyzed statically with a dumper. Also, the input program
was highly verbose (unnecessarily).

r325850 commits a clang test instead, so it's time to retire this.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@325851 91177308-0d34-0410-b5e6-96231b3b80d8