[Utility] Simplify and generalize the CleanUp helper, NFC
Removing the template arguments and most of the mutating methods from
CleanUp makes it easier to understand and reuse.

In its present state, CleanUp would be too cumbersome to adapt to cases
where multiple objects need to be released. Take for example this change
in swift-lldb:

  https://github.com/apple/swift-lldb/pull/334/files#diff-6f474df750f75c8ba675f2a8408a5629R219

This change is simple to express with the new CleanUp, but not so simple
with the old version.

Differential Revision: https://reviews.llvm.org/D43662

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@325964 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit b438a3f143199fa44630195d103d465e05425c3b)

 Conflicts:
	lldb.xcodeproj/project.pbxproj
	source/Plugins/Process/gdb-remote/ProcessGDBRemote.cpp
	unittests/Utility/CMakeLists.txt