Upstreaming avx512 register support in debugserver.  These changes were originally written by Chris Bieneman, they've undergone a number of changes since then.
Also including the debugserver bridgeos support, another arm
environment that runs Darwin akin to ios.  These codepaths are
activated when running in a bridgeos environment which we're not
set up to test today.

There's additional (small) lldb changes to handle bridgeos binaries
that still need to be merged up.

Tested on a darwin system with avx512 hardware and without.

<rdar://problem/36424951> 


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@326756 91177308-0d34-0410-b5e6-96231b3b80d8