[SwiftLanguageRuntime] Delete unused demangler functionality.
I originally thought this should live in the compiler, to then
realize, this is actually unused. Jim tells me this was needed
at some point, but it's not anymore, so let's nuke this.

It was the last bit of demangler functionality in the debugger,
now everything has been moved to the compiler or removed.

<rdar://problem/37710513>