Last batch of test-tree cleaning changes
- postmortem tests: make sure the core files are created in the build
  folder
- TestSourceManager: copy the .c file into the build dir before
  modifying it
- TestLogging: create log files in the build folder

After these changes I get a clean test run (on linux) even if I set the
source tree to be read only. It's possible some of the skipped/xfailed
tests are still creating files in the source tree, but at the moment, I
don't have plans to go hunting for those.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@328106 91177308-0d34-0410-b5e6-96231b3b80d8