[Commands] Add coverage for language swift refcount.
To the best of my knowledge this feature was not really tested.
As I'm going to do changes in this area (e.g. adding unowned
printing) I figured I could just add some coverage.

This turned out to be a nice way of experimenting writing
lldb tests in a single file (using lit). Maybe the syntax can
be cleaned up a bit, but I have to say I'm quite happy with the
ending results. Please let me know what you think.

This commit also adds the plumbing to invoke `swiftc` in lit tests,
which I didn't add before because all the swift lit tests were
using the repl.

In preparation for <rdar://30538363>.