[Swift] Simplify the extraction of the offset of a member variable.
Eliminate MemberVariableOffsetResolver, which introduced a significant
amount of overhead to allow one to delay the offset computation for a
member variable, but its only use was immediate. Instead of creating a
new Swift remote AST context for each (Swift AST context, Type) pair,
keep one Swift remote AST context per Swift AST context and use that
to immediately query the remote AST context for the member offset.

While here, replace a handcoded fetch of the "isa" pointer of a class
instance with a call to RemoteASTContext::getHeapMetadataForObject().