[cmake] Set the default test compiler to the clang built in-tree
The cmake build tries to make sure the `clang` target is available
before setting the test compiler to a clang binary built in-tree.

For reasons we don't understand, the `clang` target is not always
available. This is despite running the exact same build-script and
cmake invocations on the same bot.

As a workaround, in swift-lldb, this change unconditionally makes the
default test compiler be the clang that's built in-tree.

rdar://38935299