Simplify SwiftASTContext's "get num fields" operation.
For tuples, we can get the number of fields directly from the Swift AST.
For classes, we can use the "has superclass" AST query to simplify things.

Delete a bunch of unused code, and re-insert a commented-out break
that (if this code path were used) would have caused infinite
recursion.