NativeThreadProtocol: delete dead code
These functions were unused as everyone just went straight for the
direct operations on the register context. In fact, the
Save/RestoreAllRegisters actually appear to be wrong (inverted). Thanks
to Tatyana for pointing this out.

These functions are not very useful now that we can guarantee that each
thread always contains a valid register context, so I just delete them.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@328770 91177308-0d34-0410-b5e6-96231b3b80d8