Create one SwiftASTContext per Moudle if the scratch context fails because of irreconcilable module import errors.
When a Swift program contains several dylibs with conflicting
ClangImporter options building the common scratch SwiftASTContext
fails and the expression evaluator is unusable. This patch detects the
situation where an expression fails because of a module import error
and then switches into a mode where one scratch context per
lldb::Module is created, using only the ClangImporter flags for the
module which ideally should have been tested by the Swift compiler
that built that module. This also bifurcates the persistent expression
state, such that each lldb::Module has its own set of return
variables, but it is still a strict improvement over not having access
to the expression evaluator at all.

rdar://problem/38686915