[SwiftLanguageRuntime] Fix printing of __NSCFString.
The internal store went from being a Swift wrapped string:
Swift._NSContiguousString to just a straight up __NSCFString.

This was broken after the recent String changes, but we
didn't catch this because this test wasn't run as part
of the PR workflow. Eventually we won't need to care about
these details in lldb as everything will be hidden inside
RemoteAST, but for now, we need to fix this case.

<rdar://problem/39006341>

(cherry picked from commit 09526a948191c0d3b1a8efa363bc14797d531dea)