[debugserver] Fix LC_BUILD_VERSION load command handling.
Summary:
In one of the 2 places the LC_BUILD_VERSION load command is handled, there
is a bug preventing us from actually handling them (the address where to
read the load command was not updated). This patch factors reading the
deployment target load commands into a helper and adds testing for the 2
code paths calling the helper.

The testing is a little bit complicated because the only times those load
commands matter is when debugging a simulator process. I added a new
decorator to check that a specific SDK is available. The actual testing was
fairly easy once I knew how to run a simulated process.

Reviewers: jasonmolenda, labath

Subscribers: lldb-commits

Differential Revision: https://reviews.llvm.org/D45298

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@329374 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit cbb255f04c1d484759e01065578beb74ead9d42f)

 Conflicts:
	tools/debugserver/source/MacOSX/MachProcess.mm
	tools/debugserver/source/RNBRemote.cpp