Move OptionElementVector helper structs from Args to Options
These are not used anywhere in the Args class. They should have been
moved as a part of r327110 (Moving Option parsing from Args to Options),
but I did not notice them then.

This does not affect the layering in any way, but in makes sense for the
structs to be defined in the near the code that uses them.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@329679 91177308-0d34-0410-b5e6-96231b3b80d8