Args: replace isprint8 usage with isprint
It looks like we introduced isprint8 way back in r169417 to be used on
getopt's short_options, which we sometimes set to values which are out
of range for normal chars to indicate options with no short form.

However, this is not how the function is used in the Args class, where
we explicitly process a string character by character.

This removes the last external dependency from the Args class.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@329682 91177308-0d34-0410-b5e6-96231b3b80d8