[dotest] Use in-tree dsymutil on Darwin
Summary:
With the upstream implementation of dsymutil containing almost all
functionality from the one shipped with Xcode, we want to use the
in-tree version for running the test suite.

This will also allow us to re-enable TestUnicodeSymbols which was
failing because of the discrepancy in how Unicode symbols were hashed in
lldb and older versions of dsymutil.

Reviewers: aprantl, davide, jingham, labath

Subscribers: mgorny, llvm-commits, lldb-commits

Differential Revision: https://reviews.llvm.org/D45518

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@329889 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit a9f796f720dd6bc8f1bbb1ef356cecfce5f91833)