llgs-tests: Increase packet time out
The default packet timeout of 1 second is a bit too small for these
tests, particularly as they are working in ack-mode, which means they
need to fit twice as many packets into the timeslot.

This does not seem to be a problem on the bots, but for some people
these tests are timing out regularly. I can't be sure increasing this
will solve their problem, but this does seem like a likely culprit.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@330578 91177308-0d34-0410-b5e6-96231b3b80d8