[debugserver] Return 'ios' instead of 'iphoneos' for the ostype.
When I merged the 2 codepaths that return an OS type, I hade
checked that the places accepting 'iphoneos' would also accept
'ios', but then I got it backwards and return 'iphoneos'.

We use this value to build triples, and there 'iphoneos' is
invalid.

This also makes the test slightly simpler.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@330877 91177308-0d34-0410-b5e6-96231b3b80d8