Move the persistent variable counter into Target so it can be shared across multiple language plugins.
In a multi-language project it is counterintuitive to have a result
variables reuse numbers just because they are using a different
language plugin in LLDB (but not for example, when they are
Objective-C versus C++, since they are both handled by Clang).

This is NFC on llvm.org except for the Go plugin.

rdar://problem/39299889

Differential Revision: https://reviews.llvm.org/D46083

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@331234 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit f897241a798e9cabebef355302debc756ad67b51)